Remove logs

+++ src/modules/defence.js
@@ -2,65 +2,31 @@
 const { tools } = require('../Utils/tools');
 
 /* PARAMS */
-const version = 'v1.0.0';
-
-const allowed = `!@~\`#%$%^&*()+_=/|>.<,;:"'}]{[qwertyuiop[]';lkjhgfdsazxcvbnm1234567890-\\?><~±}{)("''".,)} `
+const version = 'v1.1';
 const ru_alphabet = 'йцукенгшщзхъфывапролджэёячсмитьбюё';
-let logs = [];
-const defenceLogo = `
-╔════════════════════════════════════════════════════════════════════╗
+const defenceLogo = `╔════════════════════════════════════════════════════════════════════╗
 ║      ██████╗░███████╗███████╗███████╗███╗░░██╗░█████╗░███████╗     ║
 ║      ██╔══██╗██╔════╝██╔════╝██╔════╝████╗░██║██╔══██╗██╔════╝     ║
 ║      ██║░░██║█████╗░░█████╗░░█████╗░░██╔██╗██║██║░░╚═╝█████╗░░     ║
 ║      ██║░░██║██╔══╝░░██╔══╝░░██╔══╝░░██║╚████║██║░░██╗██╔══╝░░     ║
 ║      ██████╔╝███████╗██║░░░░░███████╗██║░╚███║╚█████╔╝███████╗     ║
 ║      ╚═════╝░╚══════╝╚═╝░░░░░╚══════╝╚═╝░░╚══╝░╚════╝░╚══════╝     ║
-╠═══╣${version}╠════════════════════════╦════════════════════════════════╝
-║ ██╗░░░░░░█████╗░░██████╗░░██████╗ ║
-║ ██║░░░░░██╔══██╗██╔════╝░██╔════╝ ║
-║ ██║░░░░░██║░░██║██║░░██╗░╚█████╗░ ║
-║ ██║░░░░░██║░░██║██║░░╚██╗░╚═══██╗ ║
-║ ███████╗╚█████╔╝╚██████╔╝██████╔╝ ║
-║ ╚══════╝░╚════╝░░╚═════╝░╚═════╝░ ║
-╠═══════════════════════════════════╝`
+╚═══╣${version}╠═══════════════════════════════════════════════════════════╝`
 
-function createLogo() {
-    console.log(defenceLogo);
-    const nlogs = logs.reverse();
-    for (let i = 0; i < 6; i++) {
-        if (i < 5) {
-            if (nlogs[i] == null) console.log(`║`);
-            else console.log(`║ ${nlogs[i]}`);
-        } else if (i === 5) {
-            if (nlogs[i] == null) console.log(`║`);
-            else console.log(`║ ${nlogs[i]}`);
-        }
-    }
-    console.log(`╚════════════════════════════════════`)
-}
-
-setInterval(function() {
-    tools.clear();
-    createLogo();
-}, 5000)
+tools.clear();
+console.log(defenceLogo)
 
 /* CLASSES */
 class twitch {
     static run() {
         const { client } = require('./Bots/Jourloy');
-
         client.on('message', (channel, userstate, message, self) => {
             if (self) return;
             const username = userstate['username'];
-
+            if (username === 'jourloy') return;
             for (let i in message) {
                 if (ru_alphabet.includes(message[i]) === true) {
-                    if (username === 'jourloy') {
-                        logs.push(`[ MESSAGE DELETE ] >> [ TWITCH ] >> ${username} >> ${message[i]}`);
-                        return;
-                    }
                     client.deletemessage(channel, userstate['id']);
-                    logs.push(`[ MESSAGE DELETE ] >> [ DISCORD ] >> ${username}`);
                     return;
                 }
             }
@@ -78,11 +44,11 @@ class discord {
             const message = msg.content;
             const category = msg.channel.parent.name;
 
+            if (username === 'jourloy') return;
             if (category !== 'Russia' && channel.name !== 'bot') {
                 for (let i in message) {
                     if (ru_alphabet.includes(message[i]) === true) {
                         msg.delete();
-                        logs.push(`[ MESSAGE DELETE ] >> [ DISCORD ] >> ${username}`);
                         return;
                     }
                 }
